requirements.txt
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.9.7
django-appconf==1.0.2
django-imagekit==3.3
django-toolbelt==0.0.1
gunicorn==19.6.0
pilkit==1.1.13
Pillow==3.2.0
psycopg2==2.6.1
pytz==2016.4
six==1.10.0
static3==0.7.0
whitenoise==3.2
